Fix tvOS searching on every launch
